pyproject.toml
[project]
name = "ai-tinkerers-hackathon"
version = "0.1.0"
description = "AI Tinkerers x Google Cloud Hackathon Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "google-adk>=1.15.1",
    "google-genai>=0.8.0",
    "gradio>=5.47.2",
    "litellm>=1.77.5",
    "numpy>=1.24.0",
    "pyzotero>=1.6.16",
    "python-dotenv>=1.0.0",
    "streamlit>=1.50.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
]

[tool.ruff]
line-length = 119
target-version = "py311"

[tool.ruff.lint]
select = [
    "ASYNC", # flake8-async
    "B", # flake8-bugbear
    "RSE", # flake8-raise
    "F", # pyflakes
    "I", # isort
    "UP", # pyupgrade
]

[tool.ruff.format]
# Preserve existing quote style to avoid churn.
quote-style = "preserve"
